Move wasm_error implementation into a separated file and reexport it via product_common to facilitate usage in several Identity crates
