+++ bindings/wasm/iota_interaction_ts/Cargo.toml
@@ -25,8 +25,12 @@ console_error_panic_hook = { version = "0.1" }
 eyre = "0.6.12"
 fastcrypto.workspace = true
 futures = { version = "0.3" }
+iota-sdk = { version = "1.1.5", default-features = false, features = ["serde", "std"] }
 iota_interaction = { version = "0.2.1", path = "../../../iota_interaction", default-features = false }
 js-sys = { version = "0.3.61" }
+json-proof-token = { version = "0.3.5", optional = true }
+sd-jwt-payload = { version = "0.2.1", default-features = false, features = ["sha"], optional = true }
+sd-jwt-payload-rework = { package = "sd-jwt-payload", version = "0.3", features = ["sha"], optional = true }
 secret-storage = { git = "https://github.com/iotaledger/secret-storage.git", default-features = false, tag = "v0.3.0" }
 serde = { version = "1.0", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
@@ -59,3 +63,5 @@ empty_docs = "allow"
 default = []
 keytool = ["iota_interaction/keytool"]
 keypair-signer = ["iota_interaction/keypair-signer"]
+# Additional From<E> for WasmError<'_> implementations for Error types provided by additional optional dependencies.
+into-wasm-error = ["dep:sd-jwt-payload", "dep:sd-jwt-payload-rework", "dep:json-proof-token"]
