+++ product_common/src/bindings/core_client.rs
@@ -7,10 +7,10 @@ use iota_interaction::types::base_types::{IotaAddress, ObjectID, ObjectIDParseEr
 use iota_interaction::types::crypto::PublicKey;
 use iota_interaction_ts::bindings::{WasmIotaClient, WasmTransactionSigner};
 use iota_interaction_ts::core_client::{WasmCoreClient, WasmCoreClientReadOnly};
-use iota_interaction_ts::error::{Result, WasmResult};
 use iota_interaction_ts::{IotaClientAdapter, WasmPublicKey};
 use wasm_bindgen::prelude::*;
 
+use crate::bindings::wasm_error::{Result, WasmResult};
 use crate::core_client::{CoreClient, CoreClientReadOnly};
 use crate::network_name::NetworkName;
